修正： ./src/DBManager.cpp:14: error: prototype for ‘HRESULT DBManager::Running(int)’ does not match any in class ‘DBManager’ ./src/DBManager.h:11: error: candidate is: HRESULT DBManager::Running(short unsigned int)

+++ dbx/dbx_server/src/DBManager.h
@@ -8,7 +8,7 @@ class DBManager
 public:
 	DBManager();
 
-    HRESULT Running(unsigned short p_port = 3000);
+    HRESULT Running(int p_port = 3000);
     void RunOut();
 
 protected:
